fix read from file

+++ VideoEditor/Form1.cs
@@ -22,6 +22,7 @@ namespace VideoEditor
         String size;
         String start;
         String duration;
+        String fileName = "output.mp4";
         String crop;
 
         public Form1()
@@ -105,24 +106,37 @@ namespace VideoEditor
 
         private void button1_Click(object sender, EventArgs e)
         {
+            if (!Check())
+            {
+                MessageBox.Show("Убедитесь, что Все поля заполнены верно", "Ошибка",
+                        MessageBoxButtons.OK, MessageBoxIcon.Error);
+                return;
+            }
             mode = "preview";
             pictureBox1.Image = null;
             File.Delete("previewFrame.png");
             start = Start_text.Text;
             duration = Duration_text.Text;
-            if (Crop_list.SelectedItem.Equals("По правому краю"))
+            try
             {
-                crop = "right";
+                if (Crop_list.SelectedItem.Equals("По правому краю"))
+                {
+                    crop = "right";
+                }
+                else if (Crop_list.SelectedItem.Equals("По центру"))
+                {
+                    crop = "center";
+                }
+                else if (Crop_list.SelectedItem.Equals("По левому краю"))
+                {
+                    crop = "left";
+                }
+                else
+                {
+                    crop = "none";
+                }
             }
-            else if (Crop_list.SelectedItem.Equals("По центру"))
-            {
-                crop = "center";
-            }
-            else if (Crop_list.SelectedItem.Equals("По левому краю"))
-            {
-                crop = "left";
-            }
-            else {
+            catch {
                 crop = "none";
             }
             size = "auto";
@@ -136,7 +150,8 @@ namespace VideoEditor
                     pathKeyImage + "\r\n" + pos + "\r\n" + size + "\r\n" + start + "\r\n" +
                     duration + "\r\n" + crop;
             }
-            else {
+            else
+            {
                 args = mode + "\r\n" + pathVideo + "\r\n" + pathImage + "\r\n" + pathAudio + "\r\n" +
                     pathKeyImage + "\r\n" + pos + "\r\n" + size + "\r\n" + start + "\r\n" +
                     duration;
@@ -154,37 +169,15 @@ namespace VideoEditor
             {
                 pictureBox1.SizeMode = PictureBoxSizeMode.Zoom;
             }
-            else {
+            else
+            {
                 pictureBox1.SizeMode = PictureBoxSizeMode.CenterImage;
             }
         }
 
         private void RenderButton_Click(object sender, EventArgs e)
         {
-            mode = "render";
-            start = Start_text.Text;
-            duration = Duration_text.Text;
-            crop = Crop_list.SelectedItem.ToString();
-            size = "auto";
-            //String args = "preview \r\nmovie_2.mp4 \r\nboy.jpg \r\naudio.mp3 \r\nkeyingImage.jpg \r\n" +
-            //              "auto 25 26 254 180 \r\nauto \r\n139.2 \r\n3";
-            String args = mode + "\r\n" + pathVideo + "\r\n" + pathImage + "\r\n" + pathAudio + "\r\n" +
-                pathKeyImage + "\r\n" + pos + "\r\n" + size + "\r\n" + start + "\r\n" +
-                duration + "\r\n" + crop;
-            System.IO.File.WriteAllText("input.txt", args);
-            System.Diagnostics.Process proc = new System.Diagnostics.Process();
-            proc.StartInfo = new System.Diagnostics.ProcessStartInfo("script.py");
-            proc.Start();
-            pictureBox1.Image = LoadBitmap("previewFrame.png");
-            if (LoadBitmap("previewFrame.png").Height > pictureBox1.Height
-                || LoadBitmap("previewFrame.png").Width > pictureBox1.Width)
-            {
-                pictureBox1.SizeMode = PictureBoxSizeMode.Zoom;
-            }
-            else
-            {
-                pictureBox1.SizeMode = PictureBoxSizeMode.CenterImage;
-            }
+            Render();
         }
 
         private void button3_Click(object sender, EventArgs e)
@@ -204,66 +197,117 @@ namespace VideoEditor
                         // Если достигнут конец файла, прерываем считывание.
                         if (temp == null) break;
 
-                        switch (counter) {
+                        switch (counter)
+                        {
                             case 0:
                                 break;
                             case 1:
-                                pathVideo = temp;
-                                VideoLabel.Text = "Файл выбран";
+                                if (File.Exists(temp))
+                                {
+                                    pathVideo = temp;
+                                    VideoLabel.Text = "Файл выбран";
+                                }
+                                else
+                                {
+                                    MessageBox.Show("Файл с видео не найден", "Ошибка при чтении файла",
+                                    MessageBoxButtons.OK, MessageBoxIcon.Error);
+                                }
                                 break;
                             case 2:
-                                pathImage = temp;
-                                ImageLabel.Text = "Файл выбран";
+                                if (File.Exists(temp))
+                                {
+                                    pathImage = temp;
+                                    ImageLabel.Text = "Файл выбран";
+                                }
+                                else
+                                {
+                                    MessageBox.Show("Файл с изображением не найден", "Ошибка при чтении файла",
+                                    MessageBoxButtons.OK, MessageBoxIcon.Error);
+                                }
                                 break;
                             case 3:
-                                pathAudio = temp;
-                                AudioLabel.Text = "Файл выбран";
+                                if (File.Exists(temp))
+                                {
+                                    pathAudio = temp;
+                                    AudioLabel.Text = "Файл выбран";
+                                }
+                                else
+                                {
+                                    MessageBox.Show("Файл с аудио не найден", "Ошибка при чтении файла",
+                                    MessageBoxButtons.OK, MessageBoxIcon.Error);
+                                }
                                 break;
                             case 4:
-                                pathKeyImage = temp;
-                                KeyImageLabel.Text = "Файл выбран";
+                                if (File.Exists(temp))
+                                {
+                                    pathKeyImage = temp;
+                                    KeyImageLabel.Text = "Файл выбран";
+                                }
+                                else
+                                {
+                                    MessageBox.Show("Файл с областью обрезки не найден", "Ошибка при чтении файла",
+                                    MessageBoxButtons.OK, MessageBoxIcon.Error);
+                                }
                                 break;
                             case 5:
-                                temp = Regex.Replace(temp, "auto ", "");
-                                int[] array = temp.Split(' ').Select(int.Parse).ToArray();
-                                Color_R.Text = array[0].ToString();
-                                Color_G.Text = array[1].ToString();
-                                Color_B.Text = array[2].ToString();
-                                Accuracy.Text = array[3].ToString();
+                                try
+                                {
+                                    temp = Regex.Replace(temp, "auto ", "");
+                                    int[] array = temp.Split(' ').Select(int.Parse).ToArray();
+                                    Color_R.Text = array[0].ToString();
+                                    Color_G.Text = array[1].ToString();
+                                    Color_B.Text = array[2].ToString();
+                                    Accuracy.Text = array[3].ToString();
+                                }
+                                catch { }
                                 break;
                             case 6:
                                 break;
                             case 7:
-                                start = temp;
-                                Start_text.Text = temp;
+                                try
+                                {
+                                    start = temp;
+                                    Start_text.Text = temp;
+                                }
+                                catch { }
                                 break;
                             case 8:
-                                duration = temp;
-                                Duration_text.Text = temp;
+                                try
+                                {
+                                    duration = temp;
+                                    Duration_text.Text = temp;
+                                }
+                                catch { }
                                 break;
                             case 9:
-                                crop = temp;
-                                if (crop.Equals("center"))
+                                try
                                 {
-                                    Crop_list.SetSelected(1, true);
+                                    crop = temp;
+                                    if (crop.Equals("center"))
+                                    {
+                                        Crop_list.SetSelected(1, true);
+                                    }
+                                    else if (crop.Equals("right"))
+                                    {
+                                        Crop_list.SetSelected(2, true);
+                                    }
+                                    else if (crop.Equals("left"))
+                                    {
+                                        Crop_list.SetSelected(3, true);
+                                    }
+                                    else
+                                    {
+                                        Crop_list.SetSelected(0, true);
+                                    }
                                 }
-                                else if (crop.Equals("right"))
-                                {
-                                    Crop_list.SetSelected(2, true);
-                                }
-                                else if (crop.Equals("left"))
-                                {
-                                    Crop_list.SetSelected(3, true);
-                                }
-                                else {
-                                    Crop_list.SetSelected(0, true);
-                                }
+                                catch { }
                                 break;
 
                         }
                         counter += 1;
                     }
-                    if (counter < 7) {
+                    if (counter < 7)
+                    {
                         MessageBox.Show("Убедитесь, что аргументы в файле указаны верно", "Ошибка при чтении файла",
                         MessageBoxButtons.OK, MessageBoxIcon.Error);
                     }
@@ -289,6 +333,89 @@ namespace VideoEditor
             {
                 MessageBox.Show("Файл не найден", "Error!", MessageBoxButtons.OK);
                 return null;
+            }
+        }
+        public bool Check()
+        {
+            try
+            {
+                int intNumber;
+                float floatNumber;
+                intNumber = Convert.ToInt32(Color_R.Text);
+                intNumber = Convert.ToInt32(Color_G.Text);
+                intNumber = Convert.ToInt32(Color_B.Text);
+                intNumber = Convert.ToInt32(Accuracy.Text);
+                Start_text.Text = Start_text.Text.Replace('.', ',');
+                Duration_text.Text = Duration_text.Text.Replace('.', ',');
+                floatNumber = float.Parse(Start_text.Text);
+                floatNumber = float.Parse(Duration_text.Text);
+
+                Start_text.Text = Start_text.Text.Replace(',', '.');
+                Duration_text.Text = Duration_text.Text.Replace(',', '.');
+                intNumber = Crop_list.SelectedIndex;
+            }
+            catch (Exception e)
+            {
+                return false;
+            }
+            return true;
+        }
+        public void Render() {
+            if (!Check())
+            {
+                MessageBox.Show("Убедитесь, что Все поля заполнены верно", "Ошибка",
+                        MessageBoxButtons.OK, MessageBoxIcon.Error);
+                return;
+            }
+
+            mode = "render";
+            start = Start_text.Text;
+            duration = Duration_text.Text;
+
+            try
+            {
+                crop = Crop_list.SelectedItem.ToString();
+            }
+            catch {
+                crop = "none";
+            }
+            size = "auto";
+            //String args = "preview \r\nmovie_2.mp4 \r\nboy.jpg \r\naudio.mp3 \r\nkeyingImage.jpg \r\n" +
+            //              "auto 25 26 254 180 \r\nauto \r\n139.2 \r\n3";
+            String args;
+            if (crop.Equals("none"))
+            {
+                args = mode + "\r\n" + pathVideo + "\r\n" + pathImage + "\r\n" + pathAudio + "\r\n" +
+                    pathKeyImage + "\r\n" + pos + "\r\n" + size + "\r\n" + start + "\r\n" +
+                    duration + "\r\n" + fileName + "\r\n"+ crop;
+            }
+            else
+            {
+                args = mode + "\r\n" + pathVideo + "\r\n" + pathImage + "\r\n" + pathAudio + "\r\n" +
+                    pathKeyImage + "\r\n" + pos + "\r\n" + size + "\r\n" + start + "\r\n" +
+                    duration + "\r\n" + fileName;
+            }
+            System.IO.File.WriteAllText("input.txt", args);
+            System.Diagnostics.Process proc = new System.Diagnostics.Process();
+            proc.StartInfo = new System.Diagnostics.ProcessStartInfo("script.py");
+            proc.Start();
+            pictureBox1.Image = LoadBitmap("previewFrame.png");
+            if (LoadBitmap("previewFrame.png").Height > pictureBox1.Height
+                || LoadBitmap("previewFrame.png").Width > pictureBox1.Width)
+            {
+                pictureBox1.SizeMode = PictureBoxSizeMode.Zoom;
+            }
+            else
+            {
+                pictureBox1.SizeMode = PictureBoxSizeMode.CenterImage;
+            }
+        }
+        private void button4_Click(object sender, EventArgs e)
+        {
+            if (saveFileDialog1.ShowDialog() == System.Windows.Forms.DialogResult.OK)
+            {
+                fileName = saveFileDialog1.FileName;
+                Render();
             }
         }
     }
